requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
browser-cookie3==0.13.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cssselect==1.1.0
dateparser==1.1.0
demjson3==3.0.5
facebook-scraper==0.2.51
fake-useragent==0.1.11
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
jeepney==0.7.1
keyring==23.4.0
lxml==4.7.1
lz4==3.1.10
outcome==1.1.0
parse==1.19.0
pbkdf2==1.3
phantomjs==1.4.1
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.12.0
pyee==8.2.2
pyOpenSSL==21.0.0
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
requests-html==0.10.0
SecretStorage==3.3.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
wsproto==1.0.0
zipp==3.6.0
